requirements.txt
dnspython
discord.py
httpx
discord-py-slash-command
python-dotenv